+++ apps/playground/app/(oberon)/adapter.ts
@@ -3,11 +3,11 @@ import "server-only"
 import {
   authAdapterPlugin,
   databaseAdapterPlugin,
-} from "@oberoncms/adapter-turso"
+} from "@oberoncms/plugin-turso"
 import { oberonAuthPlugin } from "@oberoncms/core/auth"
 
 import { initAdapter } from "@oberoncms/core/adapter"
-import { uploadthingPlugin } from "@oberoncms/upload-thing/plugin"
+import { uploadthingPlugin } from "@oberoncms/plugin-uploadthing/plugin"
 import { sendAdapterPlugin } from "./send"
 
 export const adapter = initAdapter([