+++ recipes/nextjs/app/(oberon)/adapter.ts
@@ -3,7 +3,7 @@ import "server-only"
 import {
   authAdapterPlugin,
   databaseAdapterPlugin,
-} from "@oberoncms/adapter-turso"
+} from "@oberoncms/plugin-turso"
 import { oberonAuthPlugin } from "@oberoncms/core/auth"
 
 import { initAdapter } from "@oberoncms/core/adapter"