+++ apps/playground/CHANGELOG.md
@@ -5,4 +5,4 @@
 ### Patch Changes
 
 - Updated dependencies [e48bd75]
-  - @oberoncms/upload-thing@0.7.1
+  - @oberoncms/plugin-uploadthing@0.7.1